pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "informarl-bneck"
version = "0.1.0"
description = "InforMARL implementation for bottleneck navigation scenarios"
authors = [
    {name = "InforMARL Team"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "gymnasium",
    "matplotlib==3.10.3", 
    "numpy==2.3.1",
    "torch",
    "torch-geometric",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[project.scripts]
informarl-train = "informarl_bneck.cli.train:main"
informarl-demo = "informarl_bneck.cli.demo:main"
informarl-test = "informarl_bneck.cli.quicktest:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100